+++ extensions/disk-util/CHANGELOG.md
@@ -4,3 +4,4 @@
 
 - Adds "Quick Create Volume" command to create APFS encrypted and case-sensitive volumes.
 - Adds "Create Volume" command to create APFS encrypted and case-sensitive volumes with extra features like symlinks (more features can be added later).
+- Adds "Show Volumes" command to get details for volumes and perform operations.